refactor: reutiliza mock de benefits
