+++ packages/mocks/src/accounts.ts
@@ -1,26 +1,15 @@
-import type { Account } from "@repo/types";
+import type { Account, Benefit } from "@repo/types";
+import { benefits } from "./benefits";
 
 export const accounts: Account[] = [
   {
     id: 1,
     balance: 840,
-    benefits: {
-      id: 1,
-      name: "Alimentação",
-      color_from: "teal",
-      color_to: "blue",
-      icon: "🥦",
-    },
+    benefits: benefits[0] as Benefit,
   },
   {
     id: 2,
     balance: 360,
-    benefits: {
-      id: 2,
-      name: "Mobilidade",
-      color_from: "lime",
-      color_to: "green",
-      icon: "🚘",
-    },
+    benefits: benefits[1] as Benefit,
   },
 ];
