+++ packages/mocks/src/benefits.ts
@@ -1,0 +1,25 @@
+import { Benefit } from "@repo/types";
+
+export const benefits: Benefit[] = [
+  {
+    id: 1,
+    name: "Alimentação",
+    color_from: "teal",
+    color_to: "blue",
+    icon: "🥦",
+  },
+  {
+    id: 2,
+    name: "Mobilidade",
+    color_from: "lime",
+    color_to: "green",
+    icon: "🚘",
+  },
+  {
+    id: 3,
+    name: "Saúde",
+    icon: "🩺",
+    color_from: "#f27794",
+    color_to: "#f08779",
+  },
+];
