+++ packages/mocks/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./accounts";
+export * from "./benefits";
 export * from "./transactions";
